Make warning on forced-globalnotice message feature configurable
This introduces $wgGlobalNoticeDisplayWarningOnEditPage, which is by
default equal to false. When set to true, a warning will show to the
editor who's editing MediaWiki:Forced-globalnotice for informing
the purpose of the message.

This change also makes the areas where the message is emphasized more
semantic, by using <strong> instead of the <b> and <u> element.

See also: https://writers.stackexchange.com/q/4674

Change-Id: I020cb5cde923937d33f2ba68488dc7b8109408d1